get users by roles

+++ feature/user/src/commonMain/kotlin/dev/alxminyaev/feature/user/rds/TutorRDS.kt
@@ -3,15 +3,15 @@ package dev.alxminyaev.feature.user.rds
 import dev.alxminyaev.feature.user.model.Tutor
 
 interface TutorRDS {
-    suspend fun findById(id: Int): Tutor?
+    suspend fun findById(id: Long): Tutor?
 
-    suspend fun save(entity: Tutor): Int
+    suspend fun save(entity: Tutor): Long
 
-    suspend fun save(entities: List<Tutor>): List<Int>
+    suspend fun save(entities: List<Tutor>): List<Long>
 
     suspend fun update(entity: Tutor)
 
-    suspend fun delete(id: Int)
+    suspend fun delete(id: Long)
 
     suspend fun deleteAll()
